// frontend/src/api/models/createRecurringManualTransactionScheduleDto.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Splice API
 * Financial account management API
 * OpenAPI spec version: 1.0
 */
import type { MoneyWithSign } from './moneyWithSign'
import type { CreateRecurringManualTransactionScheduleDtoFrequency } from './createRecurringManualTransactionScheduleDtoFrequency'

export interface CreateRecurringManualTransactionScheduleDto {
  accountId: string
  amount: MoneyWithSign
  /** @minLength 1 */
  merchantName: string
  categoryId: string
  frequency?: CreateRecurringManualTransactionScheduleDtoFrequency
  /**
   * @minimum 1
   * @maximum 31
   */
  dayOfMonth: number
  /** @pattern ^\d{4}-\d{2}-\d{2}$ */
  startDate: string
  /**
   * @nullable
   * @pattern ^\d{4}-\d{2}-\d{2}$
   */
  endDate?: string | null
}