// frontend/src/api/models/updateRecurringManualTransactionScheduleDto.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Splice API
 * Financial account management API
 * OpenAPI spec version: 1.0
 */
import type { MoneyWithSign } from './moneyWithSign'

export interface UpdateRecurringManualTransactionScheduleDto {
  accountId?: string
  amount?: MoneyWithSign
  /** @minLength 1 */
  merchantName?: string
  categoryId?: string
  /**
   * @minimum 1
   * @maximum 31
   */
  dayOfMonth?: number
  /** @pattern ^\d{4}-\d{2}-\d{2}$ */
  startDate?: string
  /**
   * @nullable
   * @pattern ^\d{4}-\d{2}-\d{2}$
   */
  endDate?: string | null
  paused?: boolean
}